+++ migra/migra.py
@@ -107,6 +107,7 @@ class Migration(object):
         self.add(self.changes.tables_only_selectables())
 
         self.add(self.changes.sequences(drops_only=True))
+        self.add(self.changes.domains(drops_only=True))
         self.add(self.changes.enums(drops_only=True, modifications=False))
         self.add(self.changes.extensions(drops_only=True, modifications=False))
         self.add(self.changes.non_mv_indexes(creations_only=True))
@@ -114,6 +115,7 @@ class Migration(object):
         self.add(self.changes.non_pk_constraints(creations_only=True))
 
         self.add(self.changes.non_table_selectable_creations())
+        self.add(self.changes.domains(creations_only=True))
         self.add(self.changes.mv_indexes(creations_only=True))
 
         if privileges: