+++ migra/__init__.py
@@ -1,6 +1,6 @@
 from __future__ import unicode_literals
 
-__version__ = "1.1.0"
+__version__ = "1.2.0"
 
 from .changes import Changes
 from .command import do_command